ai-toolkit-z_image_turbo/pyproject.toml
[project]
name = "ai-toolkit"
version = "0.1.0"
description = "AI Toolkit for training and inference with diffusion models"
requires-python = ">=3.10"
dependencies = [
    "torchao==0.10.0",
    "safetensors",
    "transformers==4.57.3",
    "lycoris-lora==1.8.3",
    "flatten_json",
    "pyyaml",
    "oyaml",
    "tensorboard",
    "kornia",
    "invisible-watermark",
    "einops",
    "accelerate",
    "toml",
    "albumentations==1.4.15",
    "albucore==0.0.16",
    "pydantic",
    "omegaconf",
    "k-diffusion",
    "open_clip_torch",
    "timm",
    "prodigyopt",
    "controlnet_aux==0.0.10",
    "python-dotenv",
    "bitsandbytes",
    "hf_transfer",
    "lpips",
    "pytorch_fid",
    "optimum-quanto==0.2.4",
    "sentencepiece",
    "huggingface_hub",
    "peft",
    "gradio",
    "python-slugify",
    "opencv-python",
    "pytorch-wavelets==1.3.0",
    "matplotlib==3.10.1",
    "setuptools==69.5.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
    "ruff",
    "sphinx",
    "sphinx-rtd-theme",
    "ipython",
    "jupyter",
    "notebook",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
    "ruff",
]

[tool.black]
line-length = 120
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

